test: 테스트 분리 및 추가
