+++ app/src/androidTest/java/woowacourse/payments/OwnerTextFieldTest.kt
@@ -11,19 +11,17 @@ import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithTag
 import androidx.compose.ui.test.onNodeWithText
 import androidx.compose.ui.test.performTextInput
+import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 import woowacourse.payments.ui.component.StringTextField
 
-class StringTextFieldTest {
+class OwnerTextFieldTest {
     @get:Rule
     val composeRule = createComposeRule()
 
-    @Test
-    fun 카드소유자_이름은_30자까지만_입력이_가능하다() {
-        val long = "Thisisaverylongsentencewithoutanyspacesanditexceedsthirtycharacters"
-        val expect = long.take(30)
-
+    @Before
+    fun setup() {
         composeRule.setContent {
             var owner by remember { mutableStateOf("") }
             StringTextField(
@@ -33,6 +31,12 @@ class StringTextFieldTest {
                 maxLength = 30,
             )
         }
+    }
+
+    @Test
+    fun 카드소유자_이름은_30자까지만_입력이_가능하다() {
+        val long = "Thisisaverylongsentencewithoutanyspacesanditexceedsthirtycharacters"
+        val expect = long.take(30)
 
         // when
         composeRule.onNodeWithTag("owner").performTextInput(long)
